+++ EterTerminalDistribution/EterTerminalDistribution/ReturnOrRemoveType.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <cstdint>
 
-enum class ReturnOrRemoveCard : uint8_t {RemoveCard,ReturnCard,Default};
+enum class ReturnOrRemoveCard : uint16_t {RemoveCard,ReturnCard,Default};
